Fixed a bug
Fixed the bug that caused the app to not display the data for tab 2 if navigating to tab 3 first, or not display the data for tab 3 if navigating to tab 2 first.  This is simply done by just creating two observables, returning one for tab 2, and the new one for tab 3.  Kind of a janky way to fix it, but I wasn't sure how else to do it.